migrate tcBF to an internal plugin
