+++ src/py4d_browser_plugin/tcBF_plugin/__init__.py
@@ -1,0 +1,1 @@
+from .tcBF_plugin import tcBFPlugin
